Refactor `TabSupervisedModel` into a separate module.
Moved `TabSupervisedModel` to `abstract_model_sub.py` for better modularization. Updated relevant imports across files to reflect this change and adjusted method signatures for consistency with the new structure.
