+++ pytorch_tabnet/abstract_model.py
@@ -136,7 +136,7 @@ class TabModel(BaseEstimator):
         drop_last: bool = True,
         callbacks: Union[None, List] = None,
         pin_memory: bool = True,
-        from_unsupervised: Union[None, "TabSupervisedModel"] = None,
+        from_unsupervised: Union[None, "TabModel"] = None,
         warm_start: bool = False,
         augmentations: Union[None, Any] = None,
         compute_importance: bool = False,
@@ -365,12 +365,12 @@ class TabModel(BaseEstimator):
 
         return res_explain, res_masks
 
-    def load_weights_from_unsupervised(self, unsupervised_model: "TabSupervisedModel") -> None:
+    def load_weights_from_unsupervised(self, unsupervised_model: "TabModel") -> None:
         """Load weights from a previously trained unsupervised TabNet model.
 
         Parameters
         ----------
-        unsupervised_model : TabSupervisedModel
+        unsupervised_model : TabModel
             Previously trained unsupervised TabNet model.
 
         """
@@ -881,6 +881,3 @@ class TabModel(BaseEstimator):
         ...
 
 
-@dataclass
-class TabSupervisedModel(TabModel):
-    """Abstract base class for TabNet models."""
