+++ pytorch_tabnet/abstract_model_sub.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from pytorch_tabnet.abstract_model import TabModel
+
+
+@dataclass
+class TabSupervisedModel(TabModel):
+    """Abstract base class for TabNet models."""
