+++ pytorch_tabnet/multitask.py
@@ -9,7 +9,7 @@ import scipy
 import torch
 
 # from torch.utils.data import DataLoader
-from pytorch_tabnet.abstract_model import TabSupervisedModel
+from pytorch_tabnet.abstract_model_sub import TabSupervisedModel
 from pytorch_tabnet.data_handlers import PredictDataset, SparsePredictDataset, TBDataLoader
 from pytorch_tabnet.multiclass_utils import check_output_dim, infer_multitask_output
 from pytorch_tabnet.utils import filter_weights
